[tests] Moved test_batch_user_creation_form to base tests
